+++ src/filter-utils.js
@@ -1,0 +1,8 @@
+import FILTERS from './filters';
+import filterMarkup from './filter-markup';
+
+export const renderFilters = (selector) => {
+  FILTERS.forEach((item) => {
+    selector.innerHTML += filterMarkup(item);
+  });
+};
