+++ src/filters.js
@@ -1,33 +1,33 @@
-export const filters = [
+export default [
   {
     name: `ALL`,
-    tasksAmount: 32,
+    count: 32,
     checked: true
   },
   {
     name: `OVERDUE`,
-    tasksAmount: 1,
+    count: 1,
     disabled: true
   },
   {
     name: `TODAY`,
-    tasksAmount: 2,
+    count: 2,
     disabled: true
   },
   {
     name: `FAVORITES`,
-    tasksAmount: 3
+    count: 3
   },
   {
     name: `Repeating`,
-    tasksAmount: 5
+    count: 5
   },
   {
     name: `Tags`,
-    tasksAmount: 8
+    count: 8
   },
   {
     name: `ARCHIVE`,
-    tasksAmount: 13
+    count: 13
   }
 ];
