+++ src/utils.js
@@ -1,0 +1,6 @@
+/* Utils */
+function getRandInt(min, max) {
+  return Math.floor(min + Math.random() * (max + 1 - min));
+}
+
+export {getRandInt};
